% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgCopula.R
\name{dgCopula}
\alias{dgCopula}
\title{Random generator for multivariate survival data via copulas.}
\usage{
dgCopula(
  typeCopula = "clayton",
  theta = 1,
  typeX = "Exp",
  num1_X = 1,
  num2_X = NULL,
  typeY = "Exp",
  num1_Y = 1,
  num2_Y = NULL,
  typeCens = "None",
  num1_Cens = NULL,
  num2_Cens = NULL,
  typeSurvData = "illness-death",
  state2.prob = 0.7,
  nsim = 250
)
}
\arguments{
\item{typeCopula}{Type of copula. Possible options are \code{"clayton"}, \code{"frank"} \code{"FGM"}, 
\code{"AMH"}, \code{"gumbel-hougaard"} and \code{"joe"}. Defaults to \code{"clayton"}.}

\item{theta}{A numeric value for the space parameter.}

\item{typeX}{Type of marginal distribution. Possible options are \code{"Exp"}, \code{"Norm"} \code{"Unif"} 
and \code{"Gamma"}. Defaults to \code{"Exp"}.}

\item{num1_X}{A numeric value for the first parameter of the first marginal distribution.}

\item{num2_X}{A numeric value for the second parameter of the first marginal distribution. 
Only required for two parameter distributions.}

\item{typeY}{Type of marginal distribution. Possible options are \code{"Exp"}, \code{"Norm"},
\code{"Unif"} and \code{"Gamma"}. Defaults to \code{"Exp"}.}

\item{num1_Y}{A numeric value for the first parameter of the second marginal distribution.}

\item{num2_Y}{A numeric value for the second parameter of the second marginal distribution.
Only required for two parameter distributions.}

\item{typeCens}{Type of censuring distribution. Possible options are \code{"None"}, \code{"Unif"},
\code{"Exp"} and \code{"Wei"}. Defaults to \code{"None"}.}

\item{num1_Cens}{A numeric value for the first parameter of the censoring distribution.}

\item{num2_Cens}{A numeric value for the second parameter of the censoring distribution.
Only required for two parameter distributions.}

\item{typeSurvData}{Type of survival data. Possible options are \code{"time-to-event"}, 
\code{"recurrent"}, #' \code{"competing-risks"} and \code{"illness-death"}. Defaults to 
\code{"illness-death"}.}

\item{state2.prob}{Probability of a individual move to the intermediate state in illness-death model. 
Only required if typeSurvData =”illness-death”. Default to 0.7.}

\item{nsim}{Number of observations to be generated.}
}
\value{
A numeric vector with the random multivariate survival data. 
Meaning of the colums for the type of survival data:
Time-to-event: 
\item{T}{Surival time. T = min(Y,Z).}
\item{Z}{Censoring time variable.}
\item{Delta}{Indicator status for censoring. Delta takes 1 when Y <= Z.}

Recurrent:
\item{T1}{First gap time.}
\item{Delta1}{Censoring indicator variable for the first gap time.}
\item{T2}{Second gap time.}
\item{Delta2}{Censoring indicator variable for the second gap time.}
\item{Z}{Censoring time variable.}

Competing risks:
\item{T}{Survival time.}
\item{Z}{Censoring time variable.}
\item{Delta}{Indicator status for censoring. Delta takes 0 if the competing risk process 
does not move from the initial state at the survival time T, or the value 1 and 2 for the 
possible causes of death 1 and 2.} 

Illness-death:
\item{T1}{Sojourn time in the initial state.}
\item{Delta1}{Indicator status. Delta1 takes 1 when T1<T and T1<Z.}
\item{T}{Total survival time.}
\item{Delta}{Indicator status for censoring. Delta takes 1 when T<Z.} 
\item{Z}{Censoring time variable.}
}
\description{
This function can be used to generate multivariate 
survival data in a variety of scenarios including competing risks, recurrent event and multi-state models.
}
\examples{

sim.data<-dgCopula(typeCopula ='clayton', theta=1, typeX='Unif', num1_X=0, num2_X=5, 
                  typeY='Unif',  num1_Y=0, num2_Y=5,  typeCens='Unif', num1_Cens=0, 
                  num2_Cens=7,  nsim=250,typeSurvData='time-to-event')

head(sim.data)
            
sim.data2<-dgCopula(typeCopula ='frank', theta=10, typeX='Exp', num1_X=0.5, 
                   typeY='Exp', num1_Y=1.5, nsim=250,typeSurvData='illness-death', 
                   typeCens='Unif', num1_Cens=0, 
                   num2_Cens=4, state2.prob=0.6)
            
}
\references{
Meira-Machado, L.; de Una-Alvarez, J.; Cadarso-Suarez, C. and Andersen, P.K. Multi-state models for 
the analysis of time to event data. Statistical Methods in Medical Research, 2009, 18, 195-222.

Meira-Machado, L., Sestelo, M.; Gonlcalves, A. Nonparametric estimation of the survival function 
for ordered multivariate failure time data: A comparative study. Biometrical Journal, 2016, 58, 623-634.

L Meira-Machado, S Faria, A simulation study comparing modeling approaches in an illness-death multi-state model,
Communications in Statistics-Simulation and Computation, 2014, 43 (5), 929-946

A Moreira, J de Una-Alvarez, L Machado Presmoothing the Aalen-Johansen estimator in the illness-death model, 
Electronic Journal of Statistics, 2013, 7, 1491-1516

A Moreira, L Meira-Machado, survivalBIV: Estimation of the bivariate distribution function for sequentially
ordered events under univariate censoring, J Stat Softw, 2012, 46 (13), 1-16
}
\author{
Gustavo Soutinho, Luis Meira-Machado
}
