% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invF.R
\name{invF}
\alias{invF}
\title{Inverse of cumulative function of a marginal distribution.}
\usage{
invF(u, type = "Exp", num1 = 1, num2 = NULL)
}
\arguments{
\item{u}{A numeric value belong to the interval [0,1], corresponding to the cumulative density.}

\item{type}{Type of marginal distribution. Possible options are \code{"Exp"}, \code{"Norm"}, 
\code{"Unif"} and \code{"Gamma"}. Defaults to \code{"Exp"}.}

\item{num1}{A numeric value for the first parameter of the marginal distribution.}

\item{num2}{A numeric value for the second parameter of the marginal distribution.}
}
\value{
A numeric value corresponding to the distribution of the marginal function.
}
\description{
The value of the inverse cumulative distribution.
}
\examples{
invF(0.2)
invF(0.2,type = 'Norm', num1 = 0.2,num2 = 0.1)
}
\author{
Gustavo Soutinho, Luis Meira-Machado
}
